requirements.txt
# Core dependencies
bittensor==9.12.2
bittensor-cli==9.15.3
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
requests>=2.28.0
anthropic>=0.7.0

# Async support
asyncio-mqtt>=0.11.0
aiohttp>=3.8.0
websockets>=12.0

# Docker support
docker>=6.0.0

# Logging and monitoring
structlog>=23.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Optional: Additional LLM providers
openai>=1.0.0

# YAML support for Petri config files
pyyaml>=6.0.0

typer>=0.12.3
python-dotenv==1.2.1
aiohttp>=3.13.2
